Restore MCPClientManager test file split, add server-everything devDep
Codex's concern about parallel npx workers fighting for the npm cache
was real but had a clean root-cause fix: install
@modelcontextprotocol/server-everything as a devDep of sdk so that
`npx -y @modelcontextprotocol/server-everything` resolves to
node_modules/.bin locally instead of populating ~/.npm/_npx from the
registry. No registry contention, no cache lockfile race.

Side effect: per-spawn time drops from ~3-4s to ~0.7s, so all MCPClient
Manager tests get faster too.

With that in place, putting `server management` (~50s) and
`multiple servers` (~12s) into their own files is safe and gives back
the file-parallelism speedup that file-split was previously providing.

Local stress test: 3 back-to-back `npm test` runs, all green, 63-66s
wall time each (vs ~127s on main, ~96s on the prior commit).
